refactor(core,schemas): migrate verification records type definations
migrate verification records type definations to @logto/schema. Make them visible to all packages
