+++ packages/core/src/routes/experience/classes/verifications/new-password-identity-verification.ts
@@ -1,12 +1,10 @@
-import { type ToZodObject } from '@logto/connector-kit';
 import {
   type InteractionIdentifier,
-  interactionIdentifierGuard,
-  UsersPasswordEncryptionMethod,
+  type UsersPasswordEncryptionMethod,
   VerificationType,
+  type NewPasswordIdentityVerificationRecordData,
 } from '@logto/schemas';
 import { generateStandardId } from '@logto/shared';
-import { z } from 'zod';
 
 import RequestError from '#src/errors/RequestError/index.js';
 import type Libraries from '#src/tenants/Libraries.js';
@@ -20,25 +18,10 @@ import { interactionIdentifierToUserProfile } from '../utils.js';
 
 import { type VerificationRecord } from './verification-record.js';
 
-export type NewPasswordIdentityVerificationRecordData = {
-  id: string;
-  type: VerificationType.NewPasswordIdentity;
-  /**
-   * For now we only support username identifier for new password identity registration.
-   * For email and phone new identity registration, a `CodeVerification` record is required.
-   */
-  identifier: InteractionIdentifier;
-  passwordEncrypted?: string;
-  passwordEncryptionMethod?: UsersPasswordEncryptionMethod.Argon2i;
-};
-
-export const newPasswordIdentityVerificationRecordDataGuard = z.object({
-  id: z.string(),
-  type: z.literal(VerificationType.NewPasswordIdentity),
-  identifier: interactionIdentifierGuard,
-  passwordEncrypted: z.string().optional(),
-  passwordEncryptionMethod: z.literal(UsersPasswordEncryptionMethod.Argon2i).optional(),
-}) satisfies ToZodObject<NewPasswordIdentityVerificationRecordData>;
+export {
+  type NewPasswordIdentityVerificationRecordData,
+  newPasswordIdentityVerificationRecordDataGuard,
+} from '@logto/schemas';
 
 /**
  * NewPasswordIdentityVerification class is used for creating a new user using password + identifier.
