+++ packages/schemas/src/foundations/jsonb-types/index.ts
@@ -8,7 +8,6 @@ export * from './sentinel.js';
 export * from './users.js';
 export * from './sso-connector.js';
 export * from './applications.js';
-export * from './verification-records.js';
 export * from './account-centers.js';
 export * from './saml-application-configs.js';
 export * from './saml-application-sessions.js';
