+++ packages/schemas/src/types/verification-records/totp-verification.ts
@@ -1,0 +1,22 @@
+import { z } from 'zod';
+
+import { type ToZodObject } from '../../utils/zod.js';
+
+import { VerificationType } from './verification-type.js';
+
+export type TotpVerificationRecordData = {
+  id: string;
+  type: VerificationType.TOTP;
+  /** UserId is required for verifying or binding new TOTP */
+  userId: string;
+  secret?: string;
+  verified: boolean;
+};
+
+export const totpVerificationRecordDataGuard = z.object({
+  id: z.string(),
+  type: z.literal(VerificationType.TOTP),
+  userId: z.string(),
+  secret: z.string().optional(),
+  verified: z.boolean(),
+}) satisfies ToZodObject<TotpVerificationRecordData>;
